Fix warnings

+++ Formant Analyzer/FormantPlotter/PlotView.h
@@ -22,9 +22,9 @@
 
 @interface PlotView : UIView
 
-- (void)getData:(short int *)databuffer withLength:(int)length;
+- (void)getData:(short int *)databuffer withLength:(long)length;
 - (void)getData:(NSData *)data;
-- (void)setDisplayIdentifier:(int)displayidentifier;
+- (void)setDisplayIdentifier:(long)displayidentifier;
 
 // Write four getter functions manually for four formant frequencies.
 @property (nonatomic, readonly) double firstFFreq;
@@ -35,14 +35,14 @@
 @property (nonatomic) short int *dataBuffer; // 16 bit signed PCM
                                              // Just a pointer. Actual buffer is in audioDeviceManager (live data) or
                                              // in firstViewController (1 of 7 stored audio files).
-@property (nonatomic) int dataBufferLength;  // How many samples of the buffer needs to be processed.
-@property (nonatomic) int strongStartIdx;    // Two indices in buffer representing strong section of signal.
-@property (nonatomic) int strongEndIdx;
-@property (nonatomic) int truncatedStartIdx; // Two indices after 15% trimming of two ends
-@property (nonatomic) int truncatedEndIdx;
-@property (nonatomic) int decimatedEndIdx;   // Samples in buffer after decimation by 4.
+@property (nonatomic) long dataBufferLength;  // How many samples of the buffer needs to be processed.
+@property (nonatomic) long strongStartIdx;    // Two indices in buffer representing strong section of signal.
+@property (nonatomic) long strongEndIdx;
+@property (nonatomic) long truncatedStartIdx; // Two indices after 15% trimming of two ends
+@property (nonatomic) long truncatedEndIdx;
+@property (nonatomic) long decimatedEndIdx;   // Samples in buffer after decimation by 4.
 
-@property (nonatomic) int displayIdentifier; // What type of plot (1 out of 5) is to be displayed.
+@property (nonatomic) long displayIdentifier; // What type of plot (1 out of 5) is to be displayed.
 
 
 @end
